splitted appendonlystr into more modules
